Waiting for Model implementation. At least everything is compilable

+++ de.bwvaachen.beamoflightgame/src/main/java/de/bwvaachen/beamoflightgame/model/IBeamsOfLightPuzzleBoard.java
@@ -11,5 +11,5 @@ public interface IBeamsOfLightPuzzleBoard extends Iterable<ITile>, Serializable 
 	public ITile getTileAt(int row, int col) throws IndexOutOfBoundsException;
 	
 	public boolean isPlacementOfTileStatePossible(LightTileState state, int row, int col);
-	public ITile getTileByIndex(int index);
+	public ITile getTileByIndex(long index);
 }
